ref(ConnectionService): merge classes and move to the sdk package
